Updating the plug-in from version 2.4.0 to 2.4.1. (#36)
This commit includes all the work for v2.4.1 of Unreal Live Link. The new feature being introduced in this version is the ability to choose between Baked and Unbaked representations of the Blend Shape and Custom Attribute curves on a Character Subject (only).

The default is to leave the curves Unbaked to match the behavior of the previous release. Unbaked curves means that Live Link will attempt to recreate the shape of the curve in Unreal (match tangents, etc.).

However, for more complicated scene setups in Maya, it's not possible to match the curve in Unreal, as there's not always a simple 1:1 mapping of a curve to a blend shape weight (or custom attribute). An arbitrary number of complex nodes can be placed between any animation curve and the final weight values. To support these more complex cases, we are providing the option to Bake the curves. In this mode, the curves will be sampled at each keyframe with linear interpolation between the keys. This allows for the support of the aforementioned more complex use cases, and also adds support for Envelopes and Group Weights in the Maya Shape Editor.

Note that Baked curves send more data to Unreal but allow for more complex setups and for "live edits" of blend shapes and custom attributes, rather than only when being keyed.